Move code to `lib/`
Also rename `bin/` directory to `exe/` (there are not binaries!).
