+++ exe/flame
@@ -1,0 +1,6 @@
+#!/usr/bin/env ruby
+# frozen_string_literal: true
+
+require_relative '../lib/flame/cli'
+
+Flame::CLI.run
