+++ lib/flame/cli/new.rb
